requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
noisereduce==3.0.3
numpy==2.3.4
packaging==25.0
pillow==12.0.0
PyAudio==0.2.14
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scipy==1.16.3
six==1.17.0
tqdm==4.67.1
websocket-client==1.9.0
